terminate child processes if parent process is stopped
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C34766

+++ BusinessProcess/Code/Vanrise.BusinessProcess.Web/Controllers/BPInstanceController.cs
@@ -39,7 +39,7 @@ namespace Vanrise.BusinessProcess.Web.Controllers
 
         [HttpGet]
         [Route("GetBPInstance")]
-        public BPInstanceModel GetBPInstance(int id)
+        public BPInstanceModel GetBPInstance(long id)
         {
             BPInstanceManager manager = new BPInstanceManager();
             return BPMappers.MapInstance(manager.GetBPInstance(id));
